v1_requirements.txt
amqp==5.0.2
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
billiard==3.6.3.0
boto==2.49.0
boto3==1.16.22
botocore==1.19.22
cached-property==1.5.2
celery==5.0.2
cement==3.0.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-celery-results==2.0.0
django-chartjs==2.2.1
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-js-asset==1.2.2
django-stubs==1.7.0
django-urls==1.1.3
djangorestframework==3.12.2
djangorestframework-stubs==1.3.0
docopt==0.6.2
docutils==0.16
idna==2.10
importlib-metadata==2.0.0
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
newrelic==5.22.1.152
pathspec==0.8.1
pdfkit==0.6.1
Pillow==8.2.0
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycodestyle==2.6.0
pycrypto==2.6.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
s3transfer==0.3.3
semantic-version==2.8.5
sentry-sdk==0.19.4
six==1.15.0
sqlparse==0.4.1
stripe==2.58.0
termcolor==1.1.0
texttable==1.6.3
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
wrapt==1.12.1
XlsxWriter==1.3.7
zipp==3.4.0
